+++ src/gamelogic/rapidXML/RapidXMLAdapter.hpp
@@ -28,9 +28,9 @@ namespace GameLogic {
         void setup_document( string file_location );
 
         string get_texture_source();
-        int get_tile_height();
-        int get_tile_width();
-        int get_tile_count();
+        size_t get_tile_height();
+        size_t get_tile_width();
+        size_t get_tile_count();
         vector<vector<size_t>> get_map();
         vector<tuple<size_t, size_t, size_t, char*>> get_objects();
         string get_entity_texture( size_t id);
@@ -43,8 +43,8 @@ namespace GameLogic {
         xml_node<>* _layer_node;
         xml_node<>* _data_node;
         xml_node<>* _object_group_node;
-        int _tile_width;
-        int _tile_height;
+        size_t _tile_width;
+        size_t _tile_height;
     };
 };
 #endif // !RAPIDXMLADAPTER_HPP
